feat: add linked_schema_checks table and enhance schema check functionality
- Introduced a new table `linked_schema_checks` to manage relationships between schema checks.
- Implemented foreign key constraints to ensure referential integrity with the `schema_checks` table.
- Created indexes for efficient querying on `schema_check_id` and `linked_schema_check_id`.
- Updated migration files to reflect these changes and ensure smooth database transitions.
